updated extraction line. uses xml for 3D canvas file. hover + meta to lock valve. right click contextual menu

+++ src/extraction_line/extraction_line_canvas.py
@@ -157,8 +157,9 @@ class ExtractionLineCanvas(HasTraits):
     canvas3D = Any#Instance(ExtractionLineCanvas3DDummy)
     manager = Any
     style = Enum('2D', '3D')
-    width = Float(400)
-    height = Float(400)
+    width = Float(700)
+    height = Float(700)
+
     def __init__(self, *args, **kw):
         '''
         '''
